feat(vendor-dev): use find-foreman for eslint extraneous dependencies
This uses the findForeman package to check for dependencies in the eslint config. The benefits of this are:

- Always use the correct path in CI and locally
- Plugins can now use this configuration and all extraneous dependencies are covered in Foreman.

Right now in Katello, we do [this now](https://github.com/Katello/katello/blob/f9a239d4e42ab95dcf2561f372df00d0cb47c42b/webpack/.eslintrc.js#L36). When this PR is added, we won't need to have that config added as it will already be done from extending this config. For plugins that extend this eslint config, the foreman dependencies will be considered for the extraneous dependencies check, meaning they don't have to add them to satisfy eslint rules.

This also adds the testing packages to be checked as dependencies. This makes it so testing packages don't have to be import through @theforeman/test. With many packages available in the @theforeman/test, importing everything from @theforeman/test can obfuscate where the imports are coming from.